refactor: 디렉토리 변경 (#309)
