+++ server/momo-api/src/main/java/com/momo/config/InterceptorConfig.java
@@ -1,6 +1,6 @@
 package com.momo.config;
 
-import com.momo.domain.auth.provider.TokenProvider;
+import com.momo.domain.auth.infra.TokenProvider;
 import com.momo.interceptor.AuthorizationInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
