+++ server/momo-api/src/test/java/com/momo/common/acceptance/AcceptanceTest.java
@@ -4,9 +4,9 @@ import static com.momo.Profile.TEST;
 import static org.springframework.http.HttpHeaders.LOCATION;
 
 import com.momo.common.DatabaseCleaner;
-import com.momo.domain.auth.provider.KakaoOAuthProvider;
-import com.momo.domain.auth.provider.OAuthProviderFactory;
-import com.momo.domain.auth.provider.TokenProvider;
+import com.momo.domain.auth.domain.KakaoOAuthProvider;
+import com.momo.domain.auth.domain.OAuthProviderFactory;
+import com.momo.domain.auth.infra.TokenProvider;
 import com.momo.domain.user.entity.User;
 import com.momo.domain.user.repository.UserRepository;
 import io.restassured.RestAssured;
